Simplify specifications hierarchy
