+++ src/main/java/org/toradocu/translator/spec/Precondition.java
@@ -2,7 +2,7 @@ package org.toradocu.translator.spec;
 
 import org.toradocu.util.Checks;
 
-public class Precondition extends AbstractSpecification {
+public class Precondition extends Specification {
 
   public Precondition(Guard guard) {
     Checks.nonNullParameter(guard, "guard");
